requirements.txt
python-dotenv~=0.19.2
selenium~=4.1.0
numpy~=1.22.2
